code: Reorganize scheduler
