+++ docsrc/modules/analysis.plugin.plugin.rst
@@ -1,0 +1,7 @@
+analysis.plugin.plugin
+======================
+
+.. automodule:: analysis.plugin.plugin
+   :members:
+   :undoc-members:
+   :show-inheritance:
