+++ docsrc/modules/scheduler.analysis.plugin.rst
@@ -1,0 +1,7 @@
+scheduler.analysis.plugin
+=========================
+
+.. automodule:: scheduler.analysis.plugin
+   :members:
+   :undoc-members:
+   :show-inheritance:
