+++ docsrc/modules/scheduler.analysis.scheduler.rst
@@ -1,0 +1,7 @@
+scheduler.analysis.scheduler
+============================
+
+.. automodule:: scheduler.analysis.scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
