+++ src/scheduler/analysis/__init__.py
@@ -1,0 +1,3 @@
+from .scheduler import AnalysisScheduler  # noqa: F401
+
+ALL = ['AnalysisScheduler']
