+++ src/test/unit/scheduler/test_analysis.py
@@ -7,7 +7,8 @@ import pytest
 
 from analysis.PluginBase import AnalysisBasePlugin
 from objects.firmware import Firmware
-from scheduler.analysis import MANDATORY_PLUGINS, AnalysisScheduler
+from scheduler.analysis import AnalysisScheduler
+from scheduler.task_scheduler import MANDATORY_PLUGINS
 from test.common_helper import MockFileObject, get_test_data_dir
 from test.mock import mock_patch, mock_spy
